refactor: extract dashboard server utility helpers (issue #384)
